move route paths to library
